requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
cloudscraper==1.2.71
frozenlist==1.6.0
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.4.3
natsort==8.4.0
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
Pygments==2.19.1
pyparsing==3.2.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
soupsieve==2.6
textual==3.1.0
textual-image==0.8.2
typing_extensions==4.13.2
uc-micro-py==1.0.3
urllib3==2.4.0
yarl==1.20.0
